requirements.txt
# requirements.txt
streamlit==1.39.0
google-generativeai==0.3.2
python-dotenv==1.0.0
pycountry==23.12.11